+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.x branch
 ## 1.1 branch
 ### 1.1.0
+* `ThumbManager::clear()` and `ThumbManager::clearAll()` can return `null`;
 * updated for `php` 7.1 and `phpunit` 8.
 
 ## 1.0 branch